create abstract sql manager, log if sql fails
