+++ src/main/java/org/geysermc/connect/storage/SQLiteStorageManager.java
@@ -25,64 +25,12 @@
 
 package org.geysermc.connect.storage;
 
-import com.fasterxml.jackson.core.type.TypeReference;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.geysermc.connect.MasterServer;
-import org.geysermc.connect.utils.Player;
-import org.geysermc.connect.utils.Server;
-
-import java.io.IOException;
 import java.sql.*;
-import java.util.ArrayList;
-import java.util.List;
 
-public class SQLiteStorageManager extends AbstractStorageManager {
-
-    private final ObjectMapper mapper = new ObjectMapper();
-
-    private Connection connection;
-
+public class SQLiteStorageManager extends AbstractSQLStorageManager {
     @Override
-    public void setupStorage() {
-        try {
-            Class.forName("org.sqlite.JDBC");
-            connection = DriverManager.getConnection("jdbc:sqlite:players.db");
-
-            try (Statement createPlayersTable = connection.createStatement()) {
-                createPlayersTable.executeUpdate("CREATE TABLE IF NOT EXISTS players (xuid TEXT, servers TEXT, PRIMARY KEY(xuid));");
-            }
-        } catch (ClassNotFoundException | SQLException e) {
-            MasterServer.getInstance().getLogger().severe("Unable to load sqlite database!", e);
-        }
-    }
-
-    @Override
-    public void closeStorage() {
-        try {
-            connection.close();
-        } catch (SQLException ignored) { }
-    }
-
-    @Override
-    public void saveServers(Player player) {
-        try (Statement updatePlayersServers = connection.createStatement()) {
-            updatePlayersServers.executeUpdate("INSERT OR REPLACE INTO players(xuid, servers) VALUES('" + player.getAuthData().xuid() + "', '" + mapper.writeValueAsString(player.getServers()) + "');");
-        } catch (IOException | SQLException ignored) { }
-    }
-
-    @Override
-    public List<Server> loadServers(Player player) {
-        List<Server> servers = new ArrayList<>();
-
-        try (Statement getPlayersServers = connection.createStatement()) {
-            ResultSet rs = getPlayersServers.executeQuery("SELECT servers FROM players WHERE xuid='" + player.getAuthData().xuid() + "';");
-
-            while (rs.next()) {
-                List<Server> loadedServers = mapper.readValue(rs.getString("servers"), new TypeReference<>() {});
-                servers.addAll(loadedServers);
-            }
-        } catch (IOException | SQLException ignored) { }
-
-        return servers;
+    protected void connectToDatabase() throws ClassNotFoundException, SQLException {
+        Class.forName("org.sqlite.JDBC");
+        connection = DriverManager.getConnection("jdbc:sqlite:players.db");
     }
 }
